Visual Changes and Code Refractor
