+++ package.json
@@ -32,12 +32,14 @@
     "embla-carousel-react": "^8.3.0",
     "framer-motion": "^11.15.0",
     "input-otp": "^1.2.4",
+    "javascript-time-ago": "^2.5.11",
     "jwt-decode": "^4.0.0",
     "ldrs": "^1.1.6",
     "lucide-react": "^0.436.0",
     "motion": "^11.15.0",
     "next-themes": "^0.4.4",
     "postcss": "^8.4.41",
+    "qrcode.react": "^4.2.0",
     "react": "^19.0.0",
     "react-day-picker": "^9.5.1",
     "react-dom": "^19.0.0",
@@ -45,6 +47,7 @@
     "react-markdown": "^9.0.1",
     "react-resizable-panels": "^2.1.4",
     "react-router-dom": "^6.23.1",
+    "react-time-ago": "^7.3.3",
     "recharts": "^2.15.1",
     "sonner": "^1.7.4",
     "tailwind-merge": "^2.5.2",
