+++ src/features/referral/detail/ReferralDetailsCard.tsx
@@ -1,35 +1,33 @@
-import {ReactElement, useState} from "react";
-import {
-    Dialog,
-    DialogContent,
-    DialogDescription, DialogFooter,
-    DialogHeader,
-    DialogTitle,
-    DialogTrigger
-} from "@assets/components/shadcnui/dialog.tsx";
+import {Referral, ReferralCode} from "@/schemas/ReferralSchemas.ts";
+import {Card, CardContent, CardHeader} from "@assets/components/shadcnui/card.tsx";
+import {useReferralStore} from "@/stores/ReferralStore.ts";
+import {useEffect, useState} from "react";
 import {Button} from "@assets/components/shadcnui/button.tsx";
+import ReferralDeleteConfirmationModal
+    from "@/features/referral/modal/ReferralDeleteConfirmationModal.tsx";
+import {useForm} from "react-hook-form";
+import {
+    UpdateReferralForm,
+    UpdateReferralFormSchema
+} from "@/schemas/Forms/ReferralFormSchemas.ts";
+import {zodResolver} from "@hookform/resolvers/zod";
 import {Form, FormControl, FormField, FormItem, FormLabel, FormMessage} from "@assets/components/shadcnui/form.tsx";
 import {Input} from "@assets/components/shadcnui/input.tsx";
 import {APP_FRONTEND_FULL_PATH} from "@/constants.ts";
 import {Checkbox} from "@assets/components/shadcnui/checkbox.tsx";
-import {useForm} from "react-hook-form";
-import {CreateReferralForm, CreateReferralFormSchema} from "@/schemas/Forms/ReferralFormSchemas.ts";
-import {zodResolver} from "@hookform/resolvers/zod";
 import {ReferralStoreActionResponse} from "@/schemas/ZustandSchemas.ts";
 import {toast} from "sonner";
-import {useReferralStore} from "@/stores/ReferralStore.ts";
+import {QRCodeCanvas} from "qrcode.react";
+import ReactTimeAgo from "react-time-ago";
 
-export interface ReferralCreateDialogProps {
-    children: ReactElement| string;
-}
+export default function ReferralDetailsCard({ referralCode }: { referralCode: ReferralCode }) {
+    const {referrals, getReferralByCode, updateReferral} = useReferralStore();
+    const [referral, setReferral] = useState<Referral>();
+    const [isSaving, setIsSaving] = useState(false);
 
-export default function ReferralCreateDialog({ children }: ReferralCreateDialogProps) {
-    const {createReferral} = useReferralStore();
-    const [createReferralModalOpen, setCreateReferralModalOpen] = useState(false);
-    const [isCreating, setIsCreating] = useState(false);
-
-    const form = useForm<CreateReferralForm>({
-        resolver: zodResolver(CreateReferralFormSchema),
+    const form = useForm<UpdateReferralForm>({
+        resolver: zodResolver(UpdateReferralFormSchema),
+        values: referral,
         defaultValues: {
             label: "",
             url: "",
@@ -37,43 +35,51 @@ export default function ReferralCreateDialog({ children }: ReferralCreateDialogP
         },
     })
 
-    async function onSubmit(data: CreateReferralForm) {
-        setIsCreating(true);
-        await createReferral(data)
+    useEffect(() => {
+        (async () => {
+            const referral = await getReferralByCode(referralCode);
+            if(referral) {
+                setReferral(referral);
+                return;
+            }
+        })();
+    }, [getReferralByCode, referralCode, referrals, setReferral]);
+
+    async function onSubmit(data: UpdateReferralForm) {
+        setIsSaving(true);
+        await updateReferral(data, referralCode)
             .then((response: ReferralStoreActionResponse) => {
-                setIsCreating(false);
+                setIsSaving(false);
+
                 if(response.success) {
-                    form.reset();
-                    setCreateReferralModalOpen(false);
-                    toast("Referral created", {
+                    toast("Referral Updated", {
                         className: "",
                         description: ""
                     })
                     return;
                 }
 
-                toast("An Error occurred while create an Referral", {
+                toast("An Error occurred while edit an Referral", {
                     className: "bg-red-500",
                     description: response.message
                 })
             });
     }
 
-    return (
-        <span onClick={(event) => event.stopPropagation()}>
-            <Dialog open={createReferralModalOpen} onOpenChange={setCreateReferralModalOpen}>
-                <DialogTrigger asChild>
-                    {children}
-                </ DialogTrigger>
-                <DialogContent className="sm:max-w-[425px]">
-                    <DialogHeader>
-                        <DialogTitle>Create a referral</DialogTitle>
-                        <DialogDescription>
-                            Here you can create a new referral
-                        </DialogDescription>
-                    </DialogHeader>
+
+    return <>
+        {referral != undefined ?
+            <Card className="rounded-none">
+                <CardHeader></CardHeader>
+                <CardContent>
+                    <QRCodeCanvas value={APP_FRONTEND_FULL_PATH + "/ref/" + referral.code} />
+                    <br/>
+                    <p>Code: {referral.code}</p>
+                    <p>Created: <ReactTimeAgo date={referral.created_at} locale="en-US"/> </p>
+                    <p>Updated: <ReactTimeAgo date={referral.updated_at} locale="en-US"/></p>
+                    <br/>
                     <Form {...form}>
-                        <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8">
+                        <form className="space-y-8" onSubmit={form.handleSubmit(onSubmit)}>
                             <FormField
                                 control={form.control}
                                 name="label"
@@ -115,14 +121,16 @@ export default function ReferralCreateDialog({ children }: ReferralCreateDialogP
                                     </FormItem>
                                 )}
                             />
-                            <DialogFooter>
-                                <Button type="submit" className="bg-green-600"
-                                        disabled={isCreating}> {isCreating ? "Creating" : "Create"} </Button>
-                            </DialogFooter>
+                            <div className="flex gap-2 w-full justify-end">
+                                <ReferralDeleteConfirmationModal referral={referral}>
+                                    <Button variant="destructive" className="cursor-pointer">Delete</Button>
+                                </ReferralDeleteConfirmationModal>
+                                <Button variant="default" className="cursor-pointer" disabled={isSaving}>{isSaving ? "Saving" : "Save"}</Button>
+                            </div>
                         </form>
                     </Form>
-                </DialogContent>
-            </Dialog>
-        </span>
-    );
+                </CardContent>
+            </Card>
+        : null}
+    </>
 }
