+++ src/main.tsx
@@ -2,6 +2,10 @@ import '@/style/globals.css'
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from '@/App.tsx'
+import en from 'javascript-time-ago/locale/en'
+import TimeAgo from "javascript-time-ago";
+
+TimeAgo.addDefaultLocale(en)
 
 const root = ReactDOM.createRoot(document.getElementById('root')!);
 root.render(
