+++ src/pages/dashboard/PasteBinBoard.tsx
